refactor: 물품 데이터베이스 패키지 변경
